SDK-390 #Discard Policy "On Connection Loss" blocks Runtime on Startup
- Check for connection before blocking the runtime
- Add Unit tests
- Move SDK-389 from JmsSenderTest to AbstractSenderTest
